Fix error with multiple nodepool providers
An indentation error caused the operator to only create the last
provider listed in the nodepool config.

Additionally, add the nodepool provider to the nodepool launcher
deployment spec.

This is needed to separate the pods for different nodepool launchers
(otherwise, the deployments may consider pods for one launcher to
belong to another).

Change-Id: I76bbae948d1a633028b3565d512fb48d68111dd7