Issue a smart-reconfigure after (re-)creating zuul
Zuul no longer automatically performs a smart-reconfigure on startup
so we need to do that ourselves in case the tenant config has
changed.

There's a minor race window after the zuul CR spec changes where
the statefulset of the scheduler has not rolled out. We have to
wait for it to complete before calling smart-reconfigure, or
we risk running it on pods scheduled for deletion.

Also adding a fix from:
  https://review.opendev.org/c/zuul/zuul-operator/+/861279

This is needed to get exec in pods to work.

Change-Id: Ib35e85ed7666c2eb322971302f7f0d94a28bfa1f
Co-Authored-By: Jan Gutter <github@jangutter.com>
Co-Authored-By: Michal Nasiadka <mnasiadka@gmail.com>
Co-Authored-By: Michael Kelly <mkelly@arista.com>